refactor(Service,Repository): service interface 제거
- service interface 제거
- SpringConfig.java 제거
- Repsotiroy에 애노테이션 추가
